+++ engine-netty/src/jvmMain/kotlin/io/github/fukusaka/keel/engine/netty/NettyChannel.kt
@@ -224,9 +224,13 @@ internal class NettyChannel(
         }
     }
 
-    override fun asSource(): RawSource = ChannelSource(this, allocator)
+    @Suppress("DEPRECATION")
+    override fun asSource(): RawSource =
+        throw UnsupportedOperationException("Use asSuspendSource() instead")
 
-    override fun asSink(): RawSink = ChannelSink(this, allocator)
+    @Suppress("DEPRECATION")
+    override fun asSink(): RawSink =
+        throw UnsupportedOperationException("Use asSuspendSink() instead")
 
     override fun close() {
         if (_open) {